+++ common/src/main/java/rzk/wirelessredstone/item/SnifferItem.java
@@ -1,18 +1,19 @@
 package rzk.wirelessredstone.item;
 
+import net.minecraft.SharedConstants;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ItemUsageContext;
-import net.minecraft.nbt.NbtCompound;
+import net.minecraft.nbt.NbtElement;
 import net.minecraft.nbt.NbtHelper;
 import net.minecraft.nbt.NbtList;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.server.world.ServerWorld;
 import net.minecraft.text.ClickEvent;
 import net.minecraft.text.HoverEvent;
-import net.minecraft.text.MutableText;
 import net.minecraft.text.Text;
+import net.minecraft.text.Texts;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Formatting;
 import net.minecraft.util.Hand;
@@ -21,11 +22,13 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import rzk.wirelessredstone.WirelessRedstone;
 import rzk.wirelessredstone.ether.RedstoneEther;
+import rzk.wirelessredstone.misc.NbtKeys;
 import rzk.wirelessredstone.misc.TranslationKeys;
 import rzk.wirelessredstone.misc.WRConfig;
 import rzk.wirelessredstone.misc.WRUtils;
+import rzk.wirelessredstone.registry.ModItems;
 
-import java.util.Iterator;
+import java.util.ArrayList;
 import java.util.Set;
 
 public class SnifferItem extends FrequencyItem
@@ -35,25 +38,43 @@ public class SnifferItem extends FrequencyItem
 		super(settings);
 	}
 
-	public void setHighlightedBlocks(long timestamp, ItemStack stack, BlockPos[] coords)
+	public static BlockPos[] getHighlightedBlocks(ItemStack stack)
 	{
-		NbtCompound nbt = stack.getOrCreateNbt();
-		nbt.putLong("timestamp", timestamp);
+		if (!stack.isOf(ModItems.frequencySniffer)) return null;
 
-		NbtList list = new NbtList();
-		for (BlockPos pos : coords)
-			list.add(NbtHelper.fromBlockPos(pos));
-		nbt.put("highlights", list);
+		var nbt = stack.getNbt();
+		if (nbt == null) return null;
+
+		var list = nbt.getList(NbtKeys.HIGHLIGHTS, NbtElement.COMPOUND_TYPE);
+		if (list.isEmpty()) return null;
+
+		var coords = new BlockPos[list.size()];
+		for (int i = 0; i < list.size(); ++i)
+			coords[i] = NbtHelper.toBlockPos(list.getCompound(i));
+
+		return coords;
 	}
 
-	public void removeHighlightBlocks(ItemStack stack)
+	public static void setHighlightedBlocks(long timestamp, ItemStack stack, BlockPos[] coords)
 	{
-		if (stack.hasNbt())
-		{
-			NbtCompound nbt = stack.getNbt();
-			nbt.remove("timestamp");
-			nbt.remove("highlights");
-		}
+		if (!stack.isOf(ModItems.frequencySniffer)) return;
+
+		var nbt = stack.getOrCreateNbt();
+		nbt.putLong(NbtKeys.TIMESTAMP, timestamp);
+
+		var list = new NbtList();
+		for (var pos : coords)
+			list.add(NbtHelper.fromBlockPos(pos));
+		nbt.put(NbtKeys.HIGHLIGHTS, list);
+	}
+
+	private static void removeHighlightBlocks(ItemStack stack)
+	{
+		var nbt = stack.getNbt();
+		if (nbt == null) return;
+
+		nbt.remove("timestamp");
+		nbt.remove("highlights");
 	}
 
 	@Override
@@ -62,6 +83,34 @@ public class SnifferItem extends FrequencyItem
 		if (!context.getPlayer().isSneaking())
 			return ActionResult.PASS;
 		return super.useOnBlock(context);
+	}
+
+	private Text buildActiveTransmittersMessage(PlayerEntity player, Set<BlockPos> transmitters, Text frequencyText)
+	{
+		var texts = new ArrayList<Text>();
+		texts.add(Text.translatable(TranslationKeys.MESSAGE_TRANSMITTERS_ACTIVE, frequencyText, transmitters.size()));
+
+		for (var pos : transmitters)
+		{
+			if (texts.size() > 20)
+			{
+				texts.add(Text.literal("..."));
+				break;
+			}
+
+			var text = WRUtils.positionText(pos);
+			texts.add(text);
+
+			if (player.hasPermissionLevel(2))
+			{
+				var teleportCommand = String.format("/tp %d %d %d", pos.getX(), pos.getY() + 1, pos.getZ());
+				ClickEvent click = new ClickEvent(ClickEvent.Action.RUN_COMMAND, teleportCommand);
+				HoverEvent hover = new HoverEvent(HoverEvent.Action.SHOW_TEXT, Text.translatable(TranslationKeys.MESSAGE_TELEPORT));
+				text.setStyle(text.getStyle().withClickEvent(click).withHoverEvent(hover));
+			}
+		}
+
+		return Texts.join(texts, Text.literal("\n"));
 	}
 
 	@Override
@@ -80,67 +129,39 @@ public class SnifferItem extends FrequencyItem
 			return TypedActionResult.fail(stack);
 		}
 
-		player.getItemCooldownManager().set(this, 20);
+		player.getItemCooldownManager().set(this, SharedConstants.TICKS_PER_SECOND);
+		var result = TypedActionResult.success(stack);
+		if (world.isClient) return result;
 
-		if (!world.isClient)
+		var ether = RedstoneEther.get((ServerWorld) world);
+		if (ether == null) return result;
+
+		var transmitters = ether.getTransmitters(frequency);
+		var frequencyText = WRUtils.frequencyText(frequency);
+
+		if (transmitters.isEmpty())
 		{
-			RedstoneEther ether = RedstoneEther.get((ServerWorld) world);
-			if (ether == null) return TypedActionResult.success(stack);
-
-			Set<BlockPos> transmitters = ether.getTransmitters(frequency);
-			MutableText frequencyComponent = Text.literal(String.valueOf(frequency)).formatted(Formatting.AQUA);
-
-			if (transmitters.isEmpty())
-			{
-				player.sendMessage(Text.translatable(TranslationKeys.MESSAGE_TRANSMITTERS_EMPTY, frequencyComponent));
-				removeHighlightBlocks(stack);
-			}
-			else
-			{
-				Iterator<BlockPos> iterator = transmitters.iterator();
-				MutableText message = Text.translatable(TranslationKeys.MESSAGE_TRANSMITTERS_ACTIVE, frequencyComponent, transmitters.size());
-				message.append("\n");
-
-				while (true)
-				{
-					BlockPos transmitter = iterator.next();
-					MutableText component = Text.literal(String.format("[x: %d, y: %d, z: %d]", transmitter.getX(), transmitter.getY(), transmitter.getZ())).formatted(Formatting.YELLOW);
-
-					if (player.hasPermissionLevel(2))
-					{
-						ClickEvent click = new ClickEvent(ClickEvent.Action.RUN_COMMAND, String.format("/tp %d %d %d", transmitter.getX(), transmitter.getY() + 1, transmitter.getZ()));
-						HoverEvent hover = new HoverEvent(HoverEvent.Action.SHOW_TEXT, Text.translatable(TranslationKeys.MESSAGE_TELEPORT));
-						component.setStyle(component.getStyle().withClickEvent(click).withHoverEvent(hover));
-					}
-
-					message.append(component);
-
-					if (iterator.hasNext()) message.append("\n");
-					else break;
-
-					if (message.getString().length() >= 1000)
-					{
-						message.append("...");
-						break;
-					}
-				}
-
-				player.sendMessage(message);
-				WirelessRedstone.PLATFORM.sendSniffer((ServerPlayerEntity) player, world.getTime(), hand, transmitters.toArray(BlockPos[]::new));
-			}
+			player.sendMessage(Text.translatable(TranslationKeys.MESSAGE_TRANSMITTERS_EMPTY, frequencyText));
+			removeHighlightBlocks(stack);
+		}
+		else
+		{
+			var message = buildActiveTransmittersMessage(player, transmitters, frequencyText);
+			player.sendMessage(message);
+			WirelessRedstone.PLATFORM.sendSniffer((ServerPlayerEntity) player, world.getTime(), hand, transmitters.toArray(BlockPos[]::new));
 		}
 
-		return TypedActionResult.success(stack);
+		return result;
 	}
 
 	@Override
 	public void inventoryTick(ItemStack stack, World world, Entity entity, int slot, boolean selected)
 	{
-		if (!selected || !world.isClient || !stack.hasNbt()) return;
+		var nbt = stack.getNbt();
+		if (!selected || !world.isClient || nbt == null) return;
 
-		NbtCompound nbt = stack.getNbt();
-
-		if (world.getTime() >= nbt.getLong("timestamp") + WRConfig.highlightTimeSeconds * 20L)
+		var timeOffset = (long) WRConfig.highlightTimeSeconds * SharedConstants.TICKS_PER_SECOND;
+		if (world.getTime() >= nbt.getLong("timestamp") + timeOffset)
 			removeHighlightBlocks(stack);
 	}
 }