+++ common/src/main/java/rzk/wirelessredstone/misc/NbtKeys.java
@@ -4,4 +4,6 @@ public class NbtKeys
 {
 	public static final String FREQUENCY = "frequency";
 	public static final String CONNECTIONS = "connections";
+	public static final String HIGHLIGHTS = "highlights";
+	public static final String TIMESTAMP = "timestamp";
 }