+++ datagen/src/main/java/rzk/wirelessredstone/datagen/DefaultLanguageGenerator.java
@@ -27,8 +27,7 @@ public class DefaultLanguageGenerator extends FabricLanguageProvider
 		translations.add(ModItems.remote, "Remote");
 		translations.add(ModItems.wrench, "Wrench");
 
-		// Other
-		translations.add(TranslationKeys.ITEM_GROUP_WIRELESS_REDSTONE, "Wireless Redstone");
+		// Guis
 		translations.add(TranslationKeys.GUI_FREQUENCY_TITLE, "Frequency");
 		translations.add(TranslationKeys.GUI_CONFIG_TITLE, "Wireless Redstone Config");
 		translations.add(TranslationKeys.GUI_CONFIG_CATEGORY_GENERAL, "General");
@@ -38,12 +37,18 @@ public class DefaultLanguageGenerator extends FabricLanguageProvider
 		translations.add(TranslationKeys.GUI_CONFIG_DISPLAY_COLOR, "Display color");
 		translations.add(TranslationKeys.GUI_CONFIG_HIGHLIGHT_COLOR, "Highlight color");
 		translations.add(TranslationKeys.GUI_CONFIG_HIGHLIGHT_TIME, "Highlight time");
+
+		// Tooltips
 		translations.add(TranslationKeys.TOOLTIP_FREQUENCY, "Frequency: %s");
 		translations.add(TranslationKeys.TOOLTIP_STATE, "State: %s");
 		translations.add(TranslationKeys.TOOLTIP_STATE_OFF, "Off");
 		translations.add(TranslationKeys.TOOLTIP_STATE_ON, "On");
-		translations.add(TranslationKeys.MESSAGE_TRANSMITTERS_EMPTY, "There are no active transmitters on this frequency (%s)");
-		translations.add(TranslationKeys.MESSAGE_TRANSMITTERS_ACTIVE, "Active transmitters on this frequency (%s): %s");
+		translations.add(TranslationKeys.TOOLTIP_POSITION, "[x: %s, y: %s, z: %s]");
+
+		// Other
+		translations.add(TranslationKeys.ITEM_GROUP_WIRELESS_REDSTONE, "Wireless Redstone");
+		translations.add(TranslationKeys.MESSAGE_TRANSMITTERS_EMPTY, "No active transmitters on frequency %s");
+		translations.add(TranslationKeys.MESSAGE_TRANSMITTERS_ACTIVE, "Active transmitters on frequency %s: %s");
 		translations.add(TranslationKeys.MESSAGE_TELEPORT, "Teleport here");
 		translations.add(TranslationKeys.MESSAGE_NO_FREQUENCY, "No frequency set");
 		translations.add(TranslationKeys.NETWORKING_FAILED, "Networking error: %s");