+++ forge/src/main/java/rzk/wirelessredstone/client/WirelessRedstoneClientForge.java
@@ -5,7 +5,7 @@ import net.minecraft.util.Identifier;
 import net.minecraftforge.client.event.EntityRenderersEvent;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
-import rzk.wirelessredstone.client.render.SnifferHighlightRenderer;
+import rzk.wirelessredstone.client.render.WRWorldRendererForge;
 import rzk.wirelessredstone.registry.ModBlockEntities;
 import rzk.wirelessredstone.registry.ModItems;
 import rzk.wirelessredstone.render.RedstoneTransceiverBER;
@@ -14,7 +14,7 @@ public class WirelessRedstoneClientForge
 {
 	public static void clientSetup(FMLClientSetupEvent event)
 	{
-		MinecraftForge.EVENT_BUS.register(SnifferHighlightRenderer.class);
+		MinecraftForge.EVENT_BUS.register(WRWorldRendererForge.class);
 
 		ModelPredicateProviderRegistry.register(ModItems.remote, new Identifier("state"),
 			(stack, world, entity, seed) -> ((entity != null && stack == entity.getActiveItem()) ? 1f : 0f));