+++ neoforge/src/main/java/rzk/wirelessredstone/client/WirelessRedstoneClientNeo.java
@@ -5,7 +5,7 @@ import net.minecraft.util.Identifier;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
 import net.neoforged.neoforge.client.event.EntityRenderersEvent;
 import net.neoforged.neoforge.common.NeoForge;
-import rzk.wirelessredstone.client.render.SnifferHighlightRenderer;
+import rzk.wirelessredstone.client.render.WRWorldRendererNeo;
 import rzk.wirelessredstone.registry.ModBlockEntities;
 import rzk.wirelessredstone.registry.ModItems;
 import rzk.wirelessredstone.render.RedstoneTransceiverBER;
@@ -14,7 +14,7 @@ public class WirelessRedstoneClientNeo
 {
 	public static void clientSetup(FMLClientSetupEvent event)
 	{
-		NeoForge.EVENT_BUS.register(SnifferHighlightRenderer.class);
+		NeoForge.EVENT_BUS.register(WRWorldRendererNeo.class);
 
 		ModelPredicateProviderRegistry.register(ModItems.remote, new Identifier("state"),
 			(stack, world, entity, seed) -> ((entity != null && stack == entity.getActiveItem()) ? 1f : 0f));